@@ -24,8 +24,8 @@
   "homepage": "https://github.com/yonggwan/react-ts-boilerplate#readme",
   "dependencies": {
     "cross-env": "^7.0.2",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1"
+    "react": "18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.10.5",
@@ -48,7 +48,7 @@
     "html-loader": "^1.3.0",
     "html-webpack-plugin": "^4.4.1",
     "prettier": "^2.1.1",
-    "react-jss": "^10.4.0",
+    "react-jss": "^10.9.2",
     "source-map-loader": "^1.1.0",
     "style-loader": "^1.2.1",
     "ts-loader": "^8.0.3",
